Example DMA class and Sketch
This example Sketch and class uses the analog using DMA buffers.

It is sort of like the DMA circular class, but does not use the circular buffer, but instead uses two buffers that are chained to each other using DMA Buffers.

This is still a WIP, in particular I wish in the end to have versions that fill a buffer once and stop...