DMA operation With Timer
This adds support for Timers as well as allows the DMA buffer object to
run with just one buffer.  In this mode, the dma buffer will code will
stop the dma operation when the buffer is filled.  It then allows you to
call a method to get another buffer full of analog samples.

Added examples that use the timer object.  On Teensy 3.x boards this
will use the PDB code.  On Teensy 4, this will use unused Quad timers.

There are three examples.

The first simply uses the timers and interrupts on each analog
conversion which happens at the frequency specified.

A second one use the DMA buffer object with two buffers, and runs
continuous analog samples round robins between them and prints out
info, like min, max, average.

A third one is like the second one, except you only pass one buffer to
the dma buffer object, so it halts doing conversions when the buffer is
filled and then restarts it, when you call a method on the dma buffer
object.

All three of these should compile on Teensy 3.x and 4 boards.  I tested
a few of them on T3.6 and the second one on T3.2 as well as T4.